refactor: move variants to a separate file
